Split ddprof_cmdline into 2 files
This allows to removing dependency on ddprof::Parser for profiling lib
and some unit tests.
